Address requested changes. - Add header. - Wrap inputs in label tags. - Convert forms to SFCs in separate files. - Reorganize ResetPassword file structure. - user PropType now object. - Refactor tests.
